requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.13.5
blockbuster==1.5.25
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.2
colorclass==2.2.2
compressed-rtf==1.0.7
coverage==7.11.3
cryptography==44.0.3
dataclasses-json==0.6.7
distro==1.9.0
easygui==0.98.3
ebcdic==1.1.1
et-xmlfile==2.0.0
extract-msg==0.55.0
forbiddenfruit==0.1.4
frozenlist==1.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-tools==1.75.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib-metadata==8.7.0
independentsoft-msg==1.6
iniconfig==2.3.0
jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema-rs==0.29.1
langchain==1.0.5
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.5
langchain-openai==1.0.2
langchain-redis==0.2.4
langchain-text-splitters==1.0.0
langgraph==1.0.3
langgraph-api==0.5.13
langgraph-checkpoint==3.0.1
langgraph-checkpoint-sqlite==3.0.0
langgraph-cli==0.4.7
langgraph-prebuilt==1.0.4
langgraph-runtime-inmem==0.18.0
langgraph-sdk==0.2.9
langsmith==0.4.42
lark==1.1.9
mail-parser==4.1.4
markupsafe==3.0.3
marshmallow==3.26.1
ml-dtypes==0.5.3
msoffcrypto-tool==5.4.2
multidict==6.7.0
mypy-extensions==1.1.0
numpy==2.3.4
olefile==0.47
oletools==0.60.2
openai==2.7.2
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
pcodedmp==1.2.6
pluggy==1.6.0
ply==3.11
propcache==0.4.1
protobuf==6.33.1
psycopg==3.2.12
psycopg-binary==3.2.12
pycparser==2.23
pydantic==2.12.4
pydantic-core==2.41.5
pydantic-settings==2.12.0
pygments==2.19.2
pyjwt==2.10.1
pyparsing==3.3.0a1
pytest==9.0.1
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-ulid==3.1.0
pytz==2025.2
pyyaml==6.0.3
rapidfuzz==3.14.3
red-black-tree-mod==1.22
redis==6.4.0
redisvl==0.11.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rtfde==0.1.2.1
ruff==0.14.5
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlalchemy==2.0.44
sqlite-vec==0.1.6
sse-starlette==2.1.3
starlette==0.50.0
structlog==25.5.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
truststore==0.10.4
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
